clean up extra IO stuff
