+++ src/snowflake/snowpark/modin/core/execution/dispatching/factories/factories.py
@@ -27,26 +27,10 @@ actual implementation in the execution, bound to that factory. Each execution is
 with a Factory class.
 """
 
-import re
-import typing
-
-import pandas
+from modin.core.execution.dispatching.factories.factories import BaseFactory
 from pandas.util._decorators import doc
 
-from snowflake.snowpark.modin.core.execution.dispatching.factories.baseio import BaseIO
 from snowflake.snowpark.modin.plugin.io.snow_io import PandasOnSnowflakeIO
-
-_doc_abstract_factory_class = """
-Abstract {role} factory which allows to override the IO module easily.
-
-This class is responsible for dispatching calls of IO-functions to its
-actual execution-specific implementations.
-
-Attributes
-----------
-io_cls : BaseIO
-    IO module class of the underlying execution. The place to dispatch calls to.
-"""
 
 _doc_factory_class = """
 Factory of {execution_name} execution.
@@ -88,347 +72,18 @@ modin.pandas.{method}
 """
 )
 
-_doc_io_method_all_params = """*args : args
-    Arguments to pass to the QueryCompiler builder method.
-**kwargs : kwargs
-    Arguments to pass to the QueryCompiler builder method."""
-
 _doc_io_method_kwargs_params = """**kwargs : kwargs
     Arguments to pass to the QueryCompiler builder method."""
 
 
-types_dictionary = {"pandas": {"category": pandas.CategoricalDtype}}
-
-
-class FactoryInfo(typing.NamedTuple):
-    """
-    Structure that stores information about factory.
-
-    Parameters
-    ----------
-    engine : str
-        Name of underlying execution engine.
-    partition : str
-        Name of the partition format.
-    experimental : bool
-        Whether underlying engine is experimental-only.
-    """
-
-    engine: str
-    partition: str
-    experimental: bool
-
-
-class NotRealFactory(Exception):
-    """
-    ``NotRealFactory`` exception class.
-
-    Raise when no matching factory could be found.
-    """
-
-    pass
-
-
-@doc(_doc_abstract_factory_class, role="")
-class BaseFactory:
-    io_cls: type[BaseIO] = None  # The module where the I/O functionality exists.
-
+@doc(_doc_factory_class, backend_name="PandasOnSnowflake", execution_name="Snowflake")
+class PandasOnSnowflakeFactory(BaseFactory):
     @classmethod
-    def get_info(cls) -> FactoryInfo:
-        """
-        Get information about current factory.
-
-        Notes
-        -----
-        It parses factory name, so it must be conformant with how ``FactoryDispatcher``
-        class constructs factory names.
-        """
-        try:
-            experimental, partition, engine = re.match(
-                r"^(Experimental)?(.*)On(.*)Factory$", cls.__name__
-            ).groups()
-        except AttributeError:
-            raise NotRealFactory()
-        return FactoryInfo(
-            engine=engine, partition=partition, experimental=bool(experimental)
-        )
-
-    @classmethod
-    @doc(
-        _doc_factory_prepare_method,
-        io_module_name="an underlying execution's IO-module",
-    )
+    @doc(_doc_factory_prepare_method, io_module_name="``PandasOnSnowflakeIO``")
     def prepare(cls):
-        raise NotImplementedError("Subclasses of BaseFactory must implement prepare")
+        cls.io_cls = PandasOnSnowflakeIO
 
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="pandas DataFrame",
-        params="df : pandas.DataFrame",
-        method="utils.from_pandas",
-    )
-    def _from_pandas(cls, df):
-        return cls.io_cls.from_pandas(df)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="Arrow Table",
-        params="at : pyarrow.Table",
-        method="utils.from_arrow",
-    )
-    def _from_arrow(cls, at):
-        return cls.io_cls.from_arrow(at)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a non-pandas object (dict, list, np.array etc...)",
-        params=_doc_io_method_all_params,
-        method="utils.from_non_pandas",
-    )
-    def _from_non_pandas(cls, *args, **kwargs):
-        return cls.io_cls.from_non_pandas(*args, **kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a DataFrame object supporting exchange protocol `__dataframe__()`",
-        params=_doc_io_method_all_params,
-        method="utils.from_dataframe",
-    )
-    def _from_dataframe(cls, *args, **kwargs):
-        return cls.io_cls.from_dataframe(*args, **kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a Parquet file",
-        params=_doc_io_method_kwargs_params,
-        method="read_parquet",
-    )
-    def _read_parquet(cls, **kwargs):
-        return cls.io_cls.read_parquet(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a CSV file",
-        params=_doc_io_method_kwargs_params,
-        method="read_csv",
-    )
-    def _read_csv(cls, **kwargs):
-        return cls.io_cls.read_csv(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a JSON file",
-        params=_doc_io_method_kwargs_params,
-        method="read_json",
-    )
-    def _read_json(cls, **kwargs):
-        return cls.io_cls.read_json(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a Google BigQuery",
-        params=_doc_io_method_kwargs_params,
-        method="read_gbq",
-    )
-    def _read_gbq(cls, **kwargs):
-        return cls.io_cls.read_gbq(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="an HTML document",
-        params=_doc_io_method_kwargs_params,
-        method="read_html",
-    )
-    def _read_html(cls, **kwargs):
-        return cls.io_cls.read_html(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="clipboard",
-        params=_doc_io_method_kwargs_params,
-        method="read_clipboard",
-    )
-    def _read_clipboard(cls, **kwargs):  # pragma: no cover
-        return cls.io_cls.read_clipboard(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="an Excel file",
-        params=_doc_io_method_kwargs_params,
-        method="read_excel",
-    )
-    def _read_excel(cls, **kwargs):
-        return cls.io_cls.read_excel(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="an HDFStore",
-        params=_doc_io_method_kwargs_params,
-        method="read_hdf",
-    )
-    def _read_hdf(cls, **kwargs):
-        return cls.io_cls.read_hdf(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a feather-format object",
-        params=_doc_io_method_kwargs_params,
-        method="read_feather",
-    )
-    def _read_feather(cls, **kwargs):
-        return cls.io_cls.read_feather(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a Stata file",
-        params=_doc_io_method_kwargs_params,
-        method="read_stata",
-    )
-    def _read_stata(cls, **kwargs):
-        return cls.io_cls.read_stata(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a SAS file",
-        params=_doc_io_method_kwargs_params,
-        method="read_sas",
-    )
-    def _read_sas(cls, **kwargs):  # pragma: no cover
-        return cls.io_cls.read_sas(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a pickled Modin or pandas DataFrame",
-        params=_doc_io_method_kwargs_params,
-        method="read_pickle",
-    )
-    def _read_pickle(cls, **kwargs):
-        return cls.io_cls.read_pickle(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a SQL query or database table",
-        params=_doc_io_method_kwargs_params,
-        method="read_sql",
-    )
-    def _read_sql(cls, **kwargs):
-        return cls.io_cls.read_sql(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a table of fixed-width formatted lines",
-        params=_doc_io_method_kwargs_params,
-        method="read_fwf",
-    )
-    def _read_fwf(cls, **kwargs):
-        return cls.io_cls.read_fwf(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a SQL database table",
-        params=_doc_io_method_kwargs_params,
-        method="read_sql_table",
-    )
-    def _read_sql_table(cls, **kwargs):
-        return cls.io_cls.read_sql_table(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="a SQL query",
-        params=_doc_io_method_kwargs_params,
-        method="read_sql_query",
-    )
-    def _read_sql_query(cls, **kwargs):
-        return cls.io_cls.read_sql_query(**kwargs)
-
-    @classmethod
-    @doc(
-        _doc_io_method_template,
-        source="an SPSS file",
-        params=_doc_io_method_kwargs_params,
-        method="read_spss",
-    )
-    def _read_spss(cls, **kwargs):
-        return cls.io_cls.read_spss(**kwargs)
-
-    @classmethod
-    def _to_sql(cls, *args, **kwargs):
-        """
-        Write query compiler content to a SQL database.
-
-        Parameters
-        ----------
-        *args : args
-            Arguments to the writer method.
-        **kwargs : kwargs
-            Arguments to the writer method.
-        """
-        return cls.io_cls.to_sql(*args, **kwargs)
-
-    @classmethod
-    def _to_pickle(cls, *args, **kwargs):
-        """
-        Pickle query compiler object.
-
-        Parameters
-        ----------
-        *args : args
-            Arguments to the writer method.
-        **kwargs : kwargs
-            Arguments to the writer method.
-        """
-        return cls.io_cls.to_pickle(*args, **kwargs)
-
-    @classmethod
-    def _to_csv(cls, *args, **kwargs):
-        """
-        Write query compiler content to a CSV file.
-
-        Parameters
-        ----------
-        *args : args
-            Arguments to pass to the writer method.
-        **kwargs : kwargs
-            Arguments to pass to the writer method.
-        """
-        return cls.io_cls.to_csv(*args, **kwargs)
-
-    @classmethod
-    def _to_parquet(cls, *args, **kwargs):
-        """
-        Write query compiler content to a parquet file.
-
-        Parameters
-        ----------
-        *args : args
-            Arguments to pass to the writer method.
-        **kwargs : kwargs
-            Arguments to pass to the writer method.
-        """
-        return cls.io_cls.to_parquet(*args, **kwargs)
-
-    # following are snowflake specific functions, could also basically subclass BaseFactory into SnowflakeFactory an
-    # add the methods there.
+    # following are snowflake specific functions
     @classmethod
     @doc(
         _doc_io_method_template,
@@ -456,11 +111,3 @@ class BaseFactory:
     @classmethod
     def _to_remote(cls, *args, **kwargs):
         return cls.io_cls.to_remote(*args, **kwargs)
-
-
-@doc(_doc_factory_class, backend_name="PandasOnSnowflake", execution_name="Snowflake")
-class PandasOnSnowflakeFactory(BaseFactory):
-    @classmethod
-    @doc(_doc_factory_prepare_method, io_module_name="``PandasOnSnowflakeIO``")
-    def prepare(cls):
-        cls.io_cls = PandasOnSnowflakeIO
